Backend: Minor structure changes
