+++ Notexp-Backend/src/Models/AuthenticationResult.cs
@@ -1,11 +1,11 @@
-using System;
+using Notexp_Backend.Entities;
 
 namespace Notexp_Backend.Authorization
 {
     public class AuthenticationResult
     {
         public string AccessTokne { get; set; }
-        public string RefreshToken { get; set; }
+        public RefreshToken RefreshToken { get; set; }
         public DateTime Expiration { get; set; }
     }
 }
