+++ Notexp-Backend/src/Models/JwtSettings.cs
@@ -4,5 +4,4 @@ class JwtSettings
     string RefreshTokenSecret { get; set; }
     double AccessTokenExpirationMinutes { get; set; }
     int RefreshTokenExpirationDays { get; set; }
-
 }
