+++ Notexp-Backend/src/utils/Encoders.cs
@@ -1,10 +1,11 @@
 namespace Notexp_Backend.Utils
 {
-    class Encoders {
-        public static byte[] ToBytes (string value)
+    class Encoders
+    {
+        public static byte[] ToBytes(string value)
         {
             return System.Text.Encoding.UTF8.GetBytes(value);
-        } 
+        }
 
     }
 }
